[refactor] 字句解析の処理をファイル整理
